Cleanup -- still need to figure out how to position the indicator and not collapse the radio when selected.
